Fixed Broken build. Cleared build warnings.

+++ ext/src/Bigint.c
@@ -85,8 +85,8 @@ php_driver_bigint_init(INTERNAL_FUNCTION_PARAMETERS)
 
     if (double_value > INT64_MAX || double_value < INT64_MIN) {
       zend_throw_exception_ex(php_driver_range_exception_ce, 0 TSRMLS_CC,
-        "value must be between " LL_FORMAT " and " LL_FORMAT ", %g given",
-        INT64_MIN, INT64_MAX, double_value);
+      "value must be between " LL_FORMAT " and " LL_FORMAT ", %g given",
+      (long long int)INT64_MIN, (long long int)INT64_MAX, double_value);
       return;
     }
